feat: rewrite to ts sources.js
