+++ src/components/view/sources/sources.ts
@@ -1,0 +1,30 @@
+import './sources.css';
+
+interface SourcesItem {
+  name: string;
+  id: string;
+}
+
+class Sources {
+  draw(data: SourcesItem[]): void {
+    const fragment = document.createDocumentFragment();
+    const sourceItemTemp = document.querySelector('#sourceItemTemp') as HTMLTemplateElement;
+
+    data.forEach((item) => {
+      const sourceClone = sourceItemTemp.content.cloneNode(true) as DocumentFragment;
+
+      const sourcesItemNameElement = sourceClone.querySelector('.source__item-name') as HTMLElement;
+      sourcesItemNameElement.textContent = item.name;
+
+      const sourcesItemElement = sourceClone.querySelector('.source__item') as HTMLElement;
+      sourcesItemElement.setAttribute('data-source-id', item.id);
+
+      fragment.append(sourceClone);
+    });
+
+    const sources = document.querySelector('.sources') as HTMLElement;
+    sources.append(fragment);
+  }
+}
+
+export default Sources;
